requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
h5py==3.2.1
idna==2.10
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
lsml==0.0.1
lxml==4.6.3
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
paramiko==2.7.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
scikit-fmm==2021.2.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.3.31
urllib3==1.26.4
